+++ docs/demo/FR2/index.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import FormRender from '../../../src/antd';
 import { Button } from 'antd';
 // import FormRender from '../../../src/fusion';
@@ -9,9 +9,11 @@ const Demo = ({ schema = {} }) => {
   const [valid, setValid] = useState([]);
   const [showValid, setShowValid] = useState(true);
 
-  // useEffect(() => {
-  //   setTimeout(() => setFormData({ crowd: { upfIdList: 'sdfsdf' } }), 1500);
-  // }, []);
+  const formRef = useRef();
+
+  useEffect(() => {
+    // setTimeout(() => setFormData({ crowd: { upfIdList: 'sdfsdf' } }), 1500);
+  }, []);
 
   const onValidate = _valid => {
     console.log('没有通过的校验:', _valid);
@@ -27,18 +29,31 @@ const Demo = ({ schema = {} }) => {
     alert(JSON.stringify(formData, null, 2));
   };
 
+  const handleClick = () => {
+    formRef.current.resetData({}).then(res => {
+      console.log(res);
+    });
+  };
+
   return (
     <div style={{ maxWidth: 800 }}>
       <FormRender
+        ref={formRef}
+        readOnly={true}
         displayType="row"
         showDescIcon
         labelWidth={120}
         onValidate={onValidate}
+        onMount={() => {
+          console.log(formData);
+        }}
+        // readOnly={true}
         {...schema}
         formData={formData}
         onChange={setFormData}
         showValidate={showValid}
       />
+      <Button onClick={handleClick}>清空</Button>
       <Button type="primary" onClick={handleSubmit}>
         提交
       </Button>