+++ src/components/numberHoc.jsx
@@ -31,7 +31,7 @@ export default NumberComponent => p => {
     <NumberComponent
       {...obj}
       style={{ width: '100%', ...style }}
-      disabled={p.disabled || p.readonly}
+      disabled={p.disabled || p.readOnly}
       {...p.options}
       value={p.value}
       onChange={onChange}