+++ src/widgets/antd/index.jsx
@@ -56,5 +56,5 @@ export const mapping = {
   'range:dateTime': 'dateRange',
   '*?enum': 'select',
   'array?enum': 'checkboxes',
-  '*?readonly': 'text',
+  // '*?readOnly': 'text', // TODO: 只读模式
 };