+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ### 0.9.0
 
-- [+] 通过 ref 的方式将内部方法暴露出来，目前只用 `resetData`，具体使用参见 [api](https://x-render.gitee.io/form-render/config/api)
+- [+] 【Breaking】通过 ref 的方式将内部方法暴露出来，目前只有 `resetData`，具体使用参见 [api](https://x-render.gitee.io/form-render/config/api)
+  如果有用户在之前因为任何原因在 form-render 上挂了 ref，那这个是 breaking change(虽然感觉情况不多见)
 - [!] 修改了 onMount 的逻辑，现在 onMount 会在 formData 首次计算后触发
+- [!] 【Breaking】自定义组件接受到的入参 `readonly` 改为 `readOnly` 遵守惯例的驼峰格式。同时修复了一个 readOnly 无效的问题
+- [+] 对象类型接受 collapsed 入参，用于折叠
 
 ### 0.8.11